Fix: Rename package controllers
